Bullets firing, url loading
Many bugs but the game is getting very close to the point of basic functionality. Drawing bullets on the player 1's grid causes the whole ship to turn red, where as the other grid has red outlines of a circle. In summary, bullets are drawing but not correctly.
Linking to the visualizer with a log file url parameter will open it and begin processing immediately.